norm refactor
