+++ test.rt
@@ -1,5 +1,7 @@
-A		0.5					255,255,255
+A		0.1					255,255,255
 C	0,0,-40 		0,0,1 		50
-L	0,30,20 					1
+L	0,30,10 					1
+sp	15,10,20							25				0,200,0
 sp	15,-5,20							25				200,0,0
-pl	0,-20,0			0,1,0 						0,0,150
+cy	-20,10,70		1,1,1		10	30			200,0,200
+pl	0,-30,0			0,1,0 						0,0,150
